requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bottle==0.13.4
click==8.3.1
fastapi==0.133.1
h11==0.16.0
idna==3.11
numpy==2.4.2
pandas==3.0.1
proxy_tools==0.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pywebview==6.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==16.0
